DRILL-7160: e.q.max_rows QUERY-level option shown even if not set
The fix is to force setting to zero IFF autoLimit was intended to be set originally but is inapplicable; such as 'SHOW DATABASES'. If autolimit was not intended to be applied, setting the value to zero is not required.
WebUI is also patched to not show the zero value set in such scenarios.